style(clippy): clean up mechanical lints across the tree
This is the bulk mechanical cleanup that's been blocked on the
fmt commit landing. Now that rustfmt has a clean baseline, run
`cargo clippy --fix` and then a manual pass for the warnings
auto-fix doesn't catch.

Categories fixed:
- doc list item without indentation (26) — bare `///` lines
  after a `- item` need 2-space continuation indent
- stripping a prefix manually (7) — `path[prefix.len()..]` →
  `path.strip_prefix(prefix).unwrap()`
- if let can be collapsed into outer (5) — collapse nested
  `if let Some(x) = …` chains
- field assignment outside initializer (9) — `let mut f =
  Default::default(); f.x = v;` → `Foo { x: v, ..Default::default() }`
- &mut Vec → &mut [_] (3) — slice param signatures
- useless conversion (a few)
- sort_by_key (3) — `iter().enumerate().sort_by(…)` →
  `sort_by_key`
- clamp (1), format! in format! args (2), needless_range_loop (2)
- unused imports (a few) — mostly test helper cleanup

Categories deliberately skipped (out of scope for this commit):
- `very complex type used` (11) — needs API design decisions
  about type aliases; would change call sites
- `function has too many arguments` (5) — config struct refactor
- `is never used` / dead_code in test helpers — CI allow list
  silences these; deleting test helpers is risky
- Macro-expansion noise (clippy's known false-positive set)

Net: cargo clippy --workspace --all-targets warning count went
from 130 → 68. The CI lint job (which already passed via the
allow list) still passes with the same flags. The 68 remaining
warnings are all in the skip categories above.

Verifies:
- cargo build --tests — clean
- cargo test --lib — 811 passed, 0 failed, 1 ignored
- cargo fmt --all -- --check — clean
- cargo clippy --workspace --all-targets -- -D warnings
  -A clippy::style -A clippy::complexity -A clippy::perf
  -A clippy::pedantic -A unused -A dead_code — exit 0

+++ src/server/tools.rs
@@ -447,8 +447,8 @@ impl ToolExecutor {
                 }
             }
             "debug_sleep" => {
-                let secs = args.get("secs").and_then(|v| v.as_u64()).unwrap_or(1) as u64;
-                tokio::time::sleep(tokio::time::Duration::from_secs(secs as u64)).await;
+                let secs = args.get("secs").and_then(|v| v.as_u64()).unwrap_or(1);
+                tokio::time::sleep(tokio::time::Duration::from_secs(secs)).await;
                 return Ok(format!("Slept for {} second(s)", secs));
             }
             _ => {}
